move getProfileAPI to profile.ts
